+++ src/utils/config.ts
@@ -16,10 +16,3 @@ export const MNEMONIC = getParamOrExit('MNEMONIC');
 export const rpcUrlMainnet = getParamOrExit('RPC_URL_MAINNET');
 
 export const rpcUrlKovan = getParamOrExit('RPC_URL_KOVAN');
-
-export const SUPPORTED_NETWORKS = [
-  {chainId:     1, chainName: 'eth', name: 'Ethereum'},
-  {chainId:    42, chainName: 'kovan', name: 'Ethereum (Kovan)'},
-  {chainId:   137, chainName: 'polygon', name: 'Polygon'},
-  {chainId: 80001, chainName: 'mumbai',  name: 'Polygon (Mumbai)'},
-];
